+++ server.js
@@ -4,15 +4,18 @@ const path = require("path")
 const mongoose = require("mongoose")
 const cors = require("cors")
 const morgan = require("morgan")
+const cookieParser = require("cookie-parser")
 
 const { port, mongoURI } = require("./config.js")
 const postsRoutes = require("./routes/posts")
-const authRoutes = require("./routes/auth")
+
 
 const app = express()   // Express instance
+// Adding middlewares
 app.use(cors())
 app.use(express.json())
 app.use(morgan("tiny"))
+app.use(cookieParser)
 
 // Connecting to MongoDb database
 mongoose.connect(mongoURI, {
@@ -25,7 +28,9 @@ mongoose.connect(mongoURI, {
 
 // Routes
 app.use("/api/posts", postsRoutes)  //post route
-app.use("/api/auth", authRoutes)    //auth route
+app.use("/api/register", require("./routes/auth/register"))    //register route
+app.use("/api/login", require("./routes/auth/login"))    //login route
+app.use("/api/logout", require("./routes/auth/logout"))    //logout route
 
 app.get("/", (req, res) => {
     res.send("Homepage...")
